market/requirements.txt
Django==5.0.7
psycopg2-binary==2.9.9
python-slugify==8.0.4
Pillow==10.4.0


djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1


whitenoise==6.7.0
redis==5.0.7
celery==5.4.0


dj-database-url==2.2.0
django-cors-headers==4.4.0
drf-spectacular==0.27.2